tests/coprocessor:adjust test
