+++ tests/coprocessor/test_select.rs
@@ -10,6 +10,12 @@
 // distributed under the License is distributed on an "AS IS" BASIS,
 // See the License for the specific language governing permissions and
 // limitations under the License.
+use std::collections::{HashMap, BTreeMap};
+use std::sync::mpsc;
+use std::sync::atomic::{AtomicUsize, Ordering};
+use std::i64;
+use protobuf::{RepeatedField, Message};
+
 use tikv::coprocessor::*;
 use tikv::coprocessor;
 use kvproto::kvrpcpb::Context;
@@ -19,17 +25,12 @@ use tikv::storage::{Mutation, Key, ALL_CFS};
 use tikv::storage::engine::{self, Engine, TEMP_DIR};
 use tikv::util::worker::Worker;
 use kvproto::coprocessor::{Request, KeyRange};
-use tipb::select::{SelectRequest, SelectResponse, Chunk};
+use tipb::select::{SelectRequest, DAGRequest, SelectResponse, Chunk};
+use tipb::executor::{Executor, ExecType, TableScan, IndexScan, Selection, Aggregation, TopN, Limit};
 use tipb::schema::{self, ColumnInfo};
 use tipb::expression::{Expr, ExprType, ByItem};
 use storage::sync_storage::SyncStorage;
 use tikv::coprocessor::xeval::evaluator::FLAG_IGNORE_TRUNCATE;
-
-use std::collections::{HashMap, BTreeMap};
-use std::sync::mpsc;
-use std::sync::atomic::{AtomicUsize, Ordering};
-use std::i64;
-use protobuf::{RepeatedField, Message};
 
 static ID_GENERATOR: AtomicUsize = AtomicUsize::new(1);
 
@@ -164,6 +165,12 @@ impl Table {
     fn get_table_info(&self) -> schema::TableInfo {
         let mut tb_info = schema::TableInfo::new();
         tb_info.set_table_id(self.id);
+        tb_info.set_columns(RepeatedField::from_vec(self.get_table_columns()));
+        tb_info
+    }
+
+    fn get_table_columns(&self) -> Vec<ColumnInfo> {
+        let mut tb_info = Vec::new();
         for col in self.cols.values() {
             let mut c_info = ColumnInfo::new();
             c_info.set_column_id(col.id);
@@ -172,7 +179,7 @@ impl Table {
             if let Some(dv) = col.default_val {
                 c_info.set_default_val(datum::encode_value(&[Datum::I64(dv)]).unwrap())
             }
-            tb_info.mut_columns().push(c_info);
+            tb_info.push(c_info);
         }
         tb_info
     }
@@ -585,6 +592,221 @@ fn init_with_data(tbl: &ProductTable,
     (store, end_point)
 }
 
+fn offset_for_column(cols: &[ColumnInfo], col_id: i64) -> i64 {
+    for (offset, column) in cols.iter().enumerate() {
+        if column.get_column_id() == col_id {
+            return offset as i64;
+        }
+    }
+    0 as i64
+}
+
+struct DAGSelect {
+    execs: Vec<Executor>,
+    cols: Vec<ColumnInfo>,
+    order_by: Vec<ByItem>,
+    limit: Option<u64>,
+    aggregate: Vec<Expr>,
+    group_by: Vec<Expr>,
+    key_range: KeyRange,
+}
+
+impl DAGSelect {
+    fn from(table: &Table) -> DAGSelect {
+        let mut exec = Executor::new();
+        exec.set_tp(ExecType::TypeTableScan);
+        let mut tbl_scan = TableScan::new();
+        let mut table_info = table.get_table_info();
+        tbl_scan.set_table_id(table_info.get_table_id());
+        let columns_info = table_info.take_columns();
+        tbl_scan.set_columns(columns_info);
+        exec.set_tbl_scan(tbl_scan);
+
+        let mut range = KeyRange::new();
+        let mut buf = Vec::with_capacity(8);
+        buf.encode_i64(i64::MIN).unwrap();
+        range.set_start(table::encode_row_key(table.id, &buf));
+        buf.clear();
+        buf.encode_i64(i64::MAX).unwrap();
+        range.set_end(table::encode_row_key(table.id, &buf));
+
+
+        DAGSelect {
+            execs: vec![exec],
+            cols: table.get_table_columns(),
+            order_by: vec![],
+            limit: None,
+            aggregate: vec![],
+            group_by: vec![],
+            key_range: range,
+        }
+    }
+
+    fn from_index(table: &Table, index: Column) -> DAGSelect {
+        let idx = index.index;
+        let mut exec = Executor::new();
+        exec.set_tp(ExecType::TypeIndexScan);
+        let mut scan = IndexScan::new();
+        let mut index_info = table.get_index_info(idx);
+        scan.set_table_id(index_info.get_table_id());
+        scan.set_index_id(idx);
+
+        let columns_info = index_info.take_columns();
+        scan.set_columns(columns_info.clone());
+        exec.set_idx_scan(scan);
+
+
+        let mut range = KeyRange::new();
+
+        let mut buf = Vec::with_capacity(8);
+        buf.encode_i64(i64::MIN).unwrap();
+        range.set_start(table::encode_index_seek_key(table.id, idx, &buf));
+        buf.clear();
+        buf.encode_i64(i64::MAX).unwrap();
+        range.set_end(table::encode_index_seek_key(table.id, idx, &buf));
+
+        DAGSelect {
+            execs: vec![exec],
+            cols: columns_info.to_vec(),
+            order_by: vec![],
+            limit: None,
+            aggregate: vec![],
+            group_by: vec![],
+            key_range: range,
+        }
+    }
+
+    fn limit(mut self, n: u64) -> DAGSelect {
+        self.limit = Some(n);
+        self
+    }
+
+    fn order_by(mut self, col: Column, desc: bool) -> DAGSelect {
+        let col_offset = offset_for_column(&self.cols, col.id);
+        let mut item = ByItem::new();
+        let mut expr = Expr::new();
+        expr.set_tp(ExprType::ColumnRef);
+        expr.mut_val().encode_i64(col_offset).unwrap();
+        item.set_expr(expr);
+        item.set_desc(desc);
+        self.order_by.push(item);
+        self
+    }
+
+    fn count(mut self) -> DAGSelect {
+        let mut expr = Expr::new();
+        expr.set_tp(ExprType::Count);
+        self.aggregate.push(expr);
+        self
+    }
+
+    fn aggr_col(mut self, col: Column, aggr_t: ExprType) -> DAGSelect {
+        let col_offset = offset_for_column(&self.cols, col.id);
+        let mut col_expr = Expr::new();
+        col_expr.set_tp(ExprType::ColumnRef);
+        col_expr.mut_val().encode_i64(col_offset).unwrap();
+        let mut expr = Expr::new();
+        expr.set_tp(aggr_t);
+        expr.mut_children().push(col_expr);
+        self.aggregate.push(expr);
+        self
+    }
+
+    fn first(self, col: Column) -> DAGSelect {
+        self.aggr_col(col, ExprType::First)
+    }
+
+    fn sum(self, col: Column) -> DAGSelect {
+        self.aggr_col(col, ExprType::Sum)
+    }
+
+    fn avg(self, col: Column) -> DAGSelect {
+        self.aggr_col(col, ExprType::Avg)
+    }
+
+    fn max(self, col: Column) -> DAGSelect {
+        self.aggr_col(col, ExprType::Max)
+    }
+
+    fn min(self, col: Column) -> DAGSelect {
+        self.aggr_col(col, ExprType::Min)
+    }
+
+    fn group_by(mut self, cols: &[Column]) -> DAGSelect {
+        for col in cols {
+            let offset = offset_for_column(&self.cols, col.id);
+            let mut expr = Expr::new();
+            expr.set_tp(ExprType::ColumnRef);
+            expr.mut_val().encode_i64(offset).unwrap();
+            self.group_by.push(expr);
+        }
+        self
+    }
+
+    fn where_expr(mut self, expr: Expr) -> DAGSelect {
+        let mut exec = Executor::new();
+        exec.set_tp(ExecType::TypeSelection);
+        let mut selection = Selection::new();
+        selection.mut_conditions().push(expr);
+        exec.set_selection(selection);
+        self.execs.push(exec);
+        self
+    }
+
+    fn build(self) -> Request {
+        self.build_with(&[0])
+    }
+
+    fn build_with(mut self, flags: &[u64]) -> Request {
+        if !self.aggregate.is_empty() || !self.group_by.is_empty() {
+            let mut exec = Executor::new();
+            exec.set_tp(ExecType::TypeAggregation);
+            let mut aggr = Aggregation::new();
+            if !self.aggregate.is_empty() {
+                aggr.set_agg_func(RepeatedField::from_vec(self.aggregate));
+            }
+
+            if !self.group_by.is_empty() {
+                aggr.set_group_by(RepeatedField::from_vec(self.group_by));
+            }
+            exec.set_aggregation(aggr);
+            self.execs.push(exec);
+        }
+
+        if !self.order_by.is_empty() {
+            let mut exec = Executor::new();
+            exec.set_tp(ExecType::TypeTopN);
+            let mut topn = TopN::new();
+            topn.set_order_by(RepeatedField::from_vec(self.order_by));
+            if let Some(limit) = self.limit.take() {
+                topn.set_limit(limit);
+            }
+            exec.set_topN(topn);
+            self.execs.push(exec);
+        }
+
+        if let Some(l) = self.limit.take() {
+            let mut exec = Executor::new();
+            exec.set_tp(ExecType::TypeLimit);
+            let mut limit = Limit::new();
+            limit.set_limit(l);
+            exec.set_limit(limit);
+            self.execs.push(exec);
+        }
+
+        let mut dag = DAGRequest::new();
+        dag.set_executors(RepeatedField::from_vec(self.execs));
+        dag.set_start_ts(next_id() as u64);
+        dag.set_flags(flags.iter().fold(0, |acc, f| acc | *f));
+
+        let mut req = Request::new();
+        req.set_tp(REQ_TYPE_DAG);
+        req.set_data(dag.write_to_bytes().unwrap());
+        req.set_ranges(RepeatedField::from_vec(vec![self.key_range]));
+        req
+    }
+}
+
 #[test]
 fn test_select() {
     let data = vec![
@@ -597,16 +819,35 @@ fn test_select() {
     let product = ProductTable::new();
     let (_, mut end_point) = init_with_data(&product, &data);
 
-    let req = Select::from(&product.table).build();
-    let mut resp = handle_select(&end_point, req);
-    assert_eq!(row_cnt(resp.get_chunks()), data.len());
-    let spliter = ChunkSpliter::new(resp.take_chunks().into_vec());
-    for (row, (id, name, cnt)) in spliter.zip(data) {
-        let name_datum = name.map(|s| s.as_bytes()).into();
-        let expected_encoded = datum::encode_value(&[Datum::I64(id), name_datum, cnt.into()])
-            .unwrap();
-        assert_eq!(id, row.handle);
-        assert_eq!(row.data, &*expected_encoded);
+    // for selection
+    {
+        let req = Select::from(&product.table).build();
+        let mut resp = handle_select(&end_point, req);
+        assert_eq!(row_cnt(resp.get_chunks()), data.len());
+        let spliter = ChunkSpliter::new(resp.take_chunks().into_vec());
+        for (row, (id, name, cnt)) in spliter.zip(data.clone()) {
+            let name_datum = name.map(|s| s.as_bytes()).into();
+            let expected_encoded = datum::encode_value(&[Datum::I64(id), name_datum, cnt.into()])
+                .unwrap();
+            assert_eq!(id, row.handle);
+            assert_eq!(row.data, &*expected_encoded);
+        }
+    }
+
+    // for dag selection
+    {
+        let req = DAGSelect::from(&product.table).build();
+        let mut resp = handle_select(&end_point, req);
+        assert_eq!(row_cnt(resp.get_chunks()), data.len());
+        let spliter = ChunkSpliter::new(resp.take_chunks().into_vec());
+        for (row, (id, name, cnt)) in spliter.zip(data) {
+            let name_datum = name.map(|s| s.as_bytes()).into();
+            let expected_encoded = datum::encode_value(&[Datum::I64(id), name_datum, cnt.into()])
+                .unwrap();
+            assert_eq!(id, row.handle);
+            assert_eq!(row.data, &*expected_encoded);
+        }
+
     }
 
     end_point.stop().unwrap().join().unwrap();
@@ -623,16 +864,30 @@ fn test_group_by() {
 
     let product = ProductTable::new();
     let (_, mut end_point) = init_with_data(&product, &data);
-
-    let req = Select::from(&product.table).group_by(&[product.name]).build();
-    let mut resp = handle_select(&end_point, req);
-    // should only have name:0, name:2 and name:1
-    assert_eq!(row_cnt(resp.get_chunks()), 3);
-    let spliter = ChunkSpliter::new(resp.take_chunks().into_vec());
-    for (row, name) in spliter.zip(&[b"name:0", b"name:2", b"name:1"]) {
-        let gk = datum::encode_value(&[Datum::Bytes(name.to_vec())]).unwrap();
-        let expected_encoded = datum::encode_value(&[Datum::Bytes(gk)]).unwrap();
-        assert_eq!(row.data, &*expected_encoded);
+    // for selection
+    {
+        let req = Select::from(&product.table).group_by(&[product.name]).build();
+        let mut resp = handle_select(&end_point, req);
+        // should only have name:0, name:2 and name:1
+        assert_eq!(row_cnt(resp.get_chunks()), 3);
+        let spliter = ChunkSpliter::new(resp.take_chunks().into_vec());
+        for (row, name) in spliter.zip(&[b"name:0", b"name:2", b"name:1"]) {
+            let gk = datum::encode_value(&[Datum::Bytes(name.to_vec())]).unwrap();
+            let expected_encoded = datum::encode_value(&[Datum::Bytes(gk)]).unwrap();
+            assert_eq!(row.data, &*expected_encoded);
+        }
+    }
+    {
+        // for dag
+        let req = DAGSelect::from(&product.table).group_by(&[product.name]).build();
+        let mut resp = handle_select(&end_point, req);
+        // should only have name:0, name:2 and name:1
+        assert_eq!(row_cnt(resp.get_chunks()), 3);
+        let spliter = ChunkSpliter::new(resp.take_chunks().into_vec());
+        for (row, name) in spliter.zip(&[b"name:0", b"name:2", b"name:1"]) {
+            let expected_encoded = datum::encode_value(&[Datum::Bytes(name.to_vec())]).unwrap();
+            assert_eq!(row.data, &*expected_encoded);
+        }
     }
 
     end_point.stop().unwrap().join().unwrap();
@@ -666,15 +921,30 @@ fn test_aggr_count() {
         (Datum::Bytes(b"name:5".to_vec()), 2),
         (Datum::Null, 1),
     ];
-    let req = Select::from(&product.table).count().group_by(&[product.name]).build();
-    let mut resp = handle_select(&end_point, req);
-    assert_eq!(row_cnt(resp.get_chunks()), exp.len());
-    let spliter = ChunkSpliter::new(resp.take_chunks().into_vec());
-    for (row, (name, cnt)) in spliter.zip(exp) {
-        let gk = datum::encode_value(&[name]);
-        let expected_datum = vec![Datum::Bytes(gk.unwrap()), Datum::U64(cnt)];
-        expected_encoded = datum::encode_value(&expected_datum).unwrap();
-        assert_eq!(row.data, &*expected_encoded);
+    {
+        // for selection
+        let req = Select::from(&product.table).count().group_by(&[product.name]).build();
+        let mut resp = handle_select(&end_point, req);
+        assert_eq!(row_cnt(resp.get_chunks()), exp.len());
+        let spliter = ChunkSpliter::new(resp.take_chunks().into_vec());
+        for (row, (name, cnt)) in spliter.zip(exp.clone()) {
+            let gk = datum::encode_value(&[name]);
+            let expected_datum = vec![Datum::Bytes(gk.unwrap()), Datum::U64(cnt)];
+            expected_encoded = datum::encode_value(&expected_datum).unwrap();
+            assert_eq!(row.data, &*expected_encoded);
+        }
+    }
+    {
+        // for dag
+        let req = DAGSelect::from(&product.table).count().group_by(&[product.name]).build();
+        let mut resp = handle_select(&end_point, req);
+        assert_eq!(row_cnt(resp.get_chunks()), exp.len());
+        let spliter = ChunkSpliter::new(resp.take_chunks().into_vec());
+        for (row, (name, cnt)) in spliter.zip(exp) {
+            let expected_datum = vec![Datum::U64(cnt), name];
+            expected_encoded = datum::encode_value(&expected_datum).unwrap();
+            assert_eq!(row.data, &*expected_encoded);
+        }
     }
 
     let exp = vec![
@@ -684,15 +954,35 @@ fn test_aggr_count() {
         (vec![Datum::Bytes(b"name:5".to_vec()), Datum::I64(4)], 2),
         (vec![Datum::Null, Datum::I64(4)], 1),
     ];
-    let req = Select::from(&product.table).count().group_by(&[product.name, product.count]).build();
-    let mut resp = handle_select(&end_point, req);
-    assert_eq!(row_cnt(resp.get_chunks()), exp.len());
-    let spliter = ChunkSpliter::new(resp.take_chunks().into_vec());
-    for (row, (gk_data, cnt)) in spliter.zip(exp) {
-        let gk = datum::encode_value(&gk_data);
-        let expected_datum = vec![Datum::Bytes(gk.unwrap()), Datum::U64(cnt)];
-        expected_encoded = datum::encode_value(&expected_datum).unwrap();
-        assert_eq!(row.data, &*expected_encoded);
+    {
+        // for selection
+        let req =
+            Select::from(&product.table).count().group_by(&[product.name, product.count]).build();
+        let mut resp = handle_select(&end_point, req);
+        assert_eq!(row_cnt(resp.get_chunks()), exp.len());
+        let spliter = ChunkSpliter::new(resp.take_chunks().into_vec());
+        for (row, (gk_data, cnt)) in spliter.zip(exp.clone()) {
+            let gk = datum::encode_value(&gk_data);
+            let expected_datum = vec![Datum::Bytes(gk.unwrap()), Datum::U64(cnt)];
+            expected_encoded = datum::encode_value(&expected_datum).unwrap();
+            assert_eq!(row.data, &*expected_encoded);
+        }
+    }
+    {
+        // for dag
+        let req = DAGSelect::from(&product.table)
+            .count()
+            .group_by(&[product.name, product.count])
+            .build();
+        let mut resp = handle_select(&end_point, req);
+        assert_eq!(row_cnt(resp.get_chunks()), exp.len());
+        let spliter = ChunkSpliter::new(resp.take_chunks().into_vec());
+        for (row, (gk_data, cnt)) in spliter.zip(exp) {
+            let mut expected_datum = vec![Datum::U64(cnt)];
+            expected_datum.extend_from_slice(gk_data.as_slice());
+            expected_encoded = datum::encode_value(&expected_datum).unwrap();
+            assert_eq!(row.data, &*expected_encoded);
+        }
     }
 
     end_point.stop().unwrap().join().unwrap();
@@ -722,15 +1012,32 @@ fn test_aggr_first() {
         (Datum::Bytes(b"name:5".to_vec()), 3),
         (Datum::Null, 7),
     ];
-    let req = Select::from(&product.table).first(product.id).group_by(&[product.name]).build();
-    let mut resp = handle_select(&end_point, req);
-    assert_eq!(row_cnt(resp.get_chunks()), exp.len());
-    let spliter = ChunkSpliter::new(resp.take_chunks().into_vec());
-    for (row, (name, id)) in spliter.zip(exp) {
-        let gk = datum::encode_value(&[name]).unwrap();
-        let expected_datum = vec![Datum::Bytes(gk), Datum::I64(id)];
-        let expected_encoded = datum::encode_value(&expected_datum).unwrap();
-        assert_eq!(row.data, &*expected_encoded);
+    {
+        // for selection
+        let req = Select::from(&product.table).first(product.id).group_by(&[product.name]).build();
+        let mut resp = handle_select(&end_point, req);
+        assert_eq!(row_cnt(resp.get_chunks()), exp.len());
+        let spliter = ChunkSpliter::new(resp.take_chunks().into_vec());
+        for (row, (name, id)) in spliter.zip(exp.clone()) {
+            let gk = datum::encode_value(&[name]).unwrap();
+            let expected_datum = vec![Datum::Bytes(gk), Datum::I64(id)];
+            let expected_encoded = datum::encode_value(&expected_datum).unwrap();
+            assert_eq!(row.data, &*expected_encoded);
+        }
+    }
+    {
+        // for dag
+        let req =
+            DAGSelect::from(&product.table).first(product.id).group_by(&[product.name]).build();
+        let mut resp = handle_select(&end_point, req);
+        assert_eq!(row_cnt(resp.get_chunks()), exp.len());
+        let spliter = ChunkSpliter::new(resp.take_chunks().into_vec());
+        for (row, (name, id)) in spliter.zip(exp) {
+            let expected_datum = vec![Datum::I64(id), name];
+            let expected_encoded = datum::encode_value(&expected_datum).unwrap();
+            assert_eq!(row.data, &*expected_encoded);
+        }
+
     }
 
     let exp = vec![
@@ -741,15 +1048,32 @@ fn test_aggr_first() {
         (5, Datum::Null),
         (6, Datum::Null),
     ];
-    let req = Select::from(&product.table).first(product.name).group_by(&[product.count]).build();
-    let mut resp = handle_select(&end_point, req);
-    assert_eq!(row_cnt(resp.get_chunks()), exp.len());
-    let spliter = ChunkSpliter::new(resp.take_chunks().into_vec());
-    for (row, (count, name)) in spliter.zip(exp) {
-        let gk = datum::encode_value(&[Datum::I64(count)]).unwrap();
-        let expected_datum = vec![Datum::Bytes(gk), name];
-        let expected_encoded = datum::encode_value(&expected_datum).unwrap();
-        assert_eq!(row.data, &*expected_encoded);
+    {
+        // for selection
+        let req =
+            Select::from(&product.table).first(product.name).group_by(&[product.count]).build();
+        let mut resp = handle_select(&end_point, req);
+        assert_eq!(row_cnt(resp.get_chunks()), exp.len());
+        let spliter = ChunkSpliter::new(resp.take_chunks().into_vec());
+        for (row, (count, name)) in spliter.zip(exp.clone()) {
+            let gk = datum::encode_value(&[Datum::I64(count)]).unwrap();
+            let expected_datum = vec![Datum::Bytes(gk), name];
+            let expected_encoded = datum::encode_value(&expected_datum).unwrap();
+            assert_eq!(row.data, &*expected_encoded);
+        }
+    }
+    {
+        // for dag
+        let req =
+            DAGSelect::from(&product.table).first(product.name).group_by(&[product.count]).build();
+        let mut resp = handle_select(&end_point, req);
+        assert_eq!(row_cnt(resp.get_chunks()), exp.len());
+        let spliter = ChunkSpliter::new(resp.take_chunks().into_vec());
+        for (row, (count, name)) in spliter.zip(exp) {
+            let expected_datum = vec![name, Datum::I64(count)];
+            let expected_encoded = datum::encode_value(&expected_datum).unwrap();
+            assert_eq!(row.data, &*expected_encoded);
+        }
     }
 
     end_point.stop().unwrap().join().unwrap();
@@ -782,15 +1106,31 @@ fn test_aggr_avg() {
                    (Datum::Bytes(b"name:5".to_vec()), (Datum::Dec(8.into()), 2)),
                    (Datum::Null, (Datum::Dec(4.into()), 1)),
                    (Datum::Bytes(b"name:4".to_vec()), (Datum::Null, 0))];
-    let req = Select::from(&product.table).avg(product.count).group_by(&[product.name]).build();
-    let mut resp = handle_select(&end_point, req);
-    assert_eq!(row_cnt(resp.get_chunks()), exp.len());
-    let spliter = ChunkSpliter::new(resp.take_chunks().into_vec());
-    for (row, (name, (sum, cnt))) in spliter.zip(exp) {
-        let gk = datum::encode_value(&[name]).unwrap();
-        let expected_datum = vec![Datum::Bytes(gk), Datum::U64(cnt), sum];
-        let expected_encoded = datum::encode_value(&expected_datum).unwrap();
-        assert_eq!(row.data, &*expected_encoded);
+    {
+        // for selection
+        let req = Select::from(&product.table).avg(product.count).group_by(&[product.name]).build();
+        let mut resp = handle_select(&end_point, req);
+        assert_eq!(row_cnt(resp.get_chunks()), exp.len());
+        let spliter = ChunkSpliter::new(resp.take_chunks().into_vec());
+        for (row, (name, (sum, cnt))) in spliter.zip(exp.clone()) {
+            let gk = datum::encode_value(&[name]).unwrap();
+            let expected_datum = vec![Datum::Bytes(gk), Datum::U64(cnt), sum];
+            let expected_encoded = datum::encode_value(&expected_datum).unwrap();
+            assert_eq!(row.data, &*expected_encoded);
+        }
+    }
+    {
+        // for dag
+        let req =
+            DAGSelect::from(&product.table).avg(product.count).group_by(&[product.name]).build();
+        let mut resp = handle_select(&end_point, req);
+        assert_eq!(row_cnt(resp.get_chunks()), exp.len());
+        let spliter = ChunkSpliter::new(resp.take_chunks().into_vec());
+        for (row, (name, (sum, cnt))) in spliter.zip(exp) {
+            let expected_datum = vec![Datum::U64(cnt), sum, name];
+            let expected_encoded = datum::encode_value(&expected_datum).unwrap();
+            assert_eq!(row.data, &*expected_encoded);
+        }
     }
     end_point.stop().unwrap();
 }
@@ -815,15 +1155,31 @@ fn test_aggr_sum() {
         (Datum::Bytes(b"name:5".to_vec()), 8),
         (Datum::Null, 4),
     ];
-    let req = Select::from(&product.table).sum(product.count).group_by(&[product.name]).build();
-    let mut resp = handle_select(&end_point, req);
-    assert_eq!(row_cnt(resp.get_chunks()), exp.len());
-    let spliter = ChunkSpliter::new(resp.take_chunks().into_vec());
-    for (row, (name, cnt)) in spliter.zip(exp) {
-        let gk = datum::encode_value(&[name]).unwrap();
-        let expected_datum = vec![Datum::Bytes(gk), Datum::Dec(cnt.into())];
-        let expected_encoded = datum::encode_value(&expected_datum).unwrap();
-        assert_eq!(row.data, &*expected_encoded);
+    {
+        // for selection
+        let req = Select::from(&product.table).sum(product.count).group_by(&[product.name]).build();
+        let mut resp = handle_select(&end_point, req);
+        assert_eq!(row_cnt(resp.get_chunks()), exp.len());
+        let spliter = ChunkSpliter::new(resp.take_chunks().into_vec());
+        for (row, (name, cnt)) in spliter.zip(exp.clone()) {
+            let gk = datum::encode_value(&[name]).unwrap();
+            let expected_datum = vec![Datum::Bytes(gk), Datum::Dec(cnt.into())];
+            let expected_encoded = datum::encode_value(&expected_datum).unwrap();
+            assert_eq!(row.data, &*expected_encoded);
+        }
+    }
+    {
+        // for dag
+        let req =
+            DAGSelect::from(&product.table).sum(product.count).group_by(&[product.name]).build();
+        let mut resp = handle_select(&end_point, req);
+        assert_eq!(row_cnt(resp.get_chunks()), exp.len());
+        let spliter = ChunkSpliter::new(resp.take_chunks().into_vec());
+        for (row, (name, cnt)) in spliter.zip(exp) {
+            let expected_datum = vec![Datum::Dec(cnt.into()), name];
+            let expected_encoded = datum::encode_value(&expected_datum).unwrap();
+            assert_eq!(row.data, &*expected_encoded);
+        }
     }
     end_point.stop().unwrap();
 }
@@ -859,20 +1215,41 @@ fn test_aggr_extre() {
         (Datum::Null, Datum::I64(4), Datum::I64(4)),
         (Datum::Bytes(b"name:6".to_vec()), Datum::Null, Datum::Null),
     ];
-    let req = Select::from(&product.table)
-        .max(product.count)
-        .min(product.count)
-        .group_by(&[product.name])
-        .build();
-    let mut resp = handle_select(&end_point, req);
-    assert_eq!(row_cnt(resp.get_chunks()), exp.len());
-    let spliter = ChunkSpliter::new(resp.take_chunks().into_vec());
-    for (row, (name, max, min)) in spliter.zip(exp) {
-        let gk = datum::encode_value(&[name]).unwrap();
-        let expected_datum = vec![Datum::Bytes(gk), max, min];
-        let expected_encoded = datum::encode_value(&expected_datum).unwrap();
-        assert_eq!(row.data, &*expected_encoded);
+    {
+        // for selection
+        let req = Select::from(&product.table)
+            .max(product.count)
+            .min(product.count)
+            .group_by(&[product.name])
+            .build();
+        let mut resp = handle_select(&end_point, req);
+        assert_eq!(row_cnt(resp.get_chunks()), exp.len());
+        let spliter = ChunkSpliter::new(resp.take_chunks().into_vec());
+        for (row, (name, max, min)) in spliter.zip(exp.clone()) {
+            let gk = datum::encode_value(&[name]).unwrap();
+            let expected_datum = vec![Datum::Bytes(gk), max, min];
+            let expected_encoded = datum::encode_value(&expected_datum).unwrap();
+            assert_eq!(row.data, &*expected_encoded);
+        }
     }
+
+    {
+        // for dag
+        let req = DAGSelect::from(&product.table)
+            .max(product.count)
+            .min(product.count)
+            .group_by(&[product.name])
+            .build();
+        let mut resp = handle_select(&end_point, req);
+        assert_eq!(row_cnt(resp.get_chunks()), exp.len());
+        let spliter = ChunkSpliter::new(resp.take_chunks().into_vec());
+        for (row, (name, max, min)) in spliter.zip(exp) {
+            let expected_datum = vec![max, min, name];
+            let expected_encoded = datum::encode_value(&expected_datum).unwrap();
+            assert_eq!(row.data, &*expected_encoded);
+        }
+    }
+
     end_point.stop().unwrap();
 }
 
@@ -898,20 +1275,43 @@ fn test_order_by_column() {
 
     let product = ProductTable::new();
     let (_, mut end_point) = init_with_data(&product, &data);
-    let req = Select::from(&product.table)
-        .order_by(product.count, true)
-        .order_by(product.name, false)
-        .limit(5)
-        .build();
-    let mut resp = handle_select(&end_point, req);
-    assert_eq!(row_cnt(resp.get_chunks()), 5);
-    let spliter = ChunkSpliter::new(resp.take_chunks().into_vec());
-    for (row, (id, name, cnt)) in spliter.zip(exp) {
-        let name_datum = name.map(|s| s.as_bytes()).into();
-        let expected_encoded =
-            datum::encode_value(&[(id as i64).into(), name_datum, (cnt as i64).into()]).unwrap();
-        assert_eq!(id as i64, row.handle);
-        assert_eq!(row.data, &*expected_encoded);
+    {
+        // for selection
+        let req = Select::from(&product.table)
+            .order_by(product.count, true)
+            .order_by(product.name, false)
+            .limit(5)
+            .build();
+        let mut resp = handle_select(&end_point, req);
+        assert_eq!(row_cnt(resp.get_chunks()), 5);
+        let spliter = ChunkSpliter::new(resp.take_chunks().into_vec());
+        for (row, (id, name, cnt)) in spliter.zip(exp.clone()) {
+            let name_datum = name.map(|s| s.as_bytes()).into();
+            let expected_encoded =
+                datum::encode_value(&[(id as i64).into(), name_datum, (cnt as i64).into()])
+                    .unwrap();
+            assert_eq!(id as i64, row.handle);
+            assert_eq!(row.data, &*expected_encoded);
+        }
+    }
+    {
+        // for dag
+        let req = DAGSelect::from(&product.table)
+            .order_by(product.count, true)
+            .order_by(product.name, false)
+            .limit(5)
+            .build();
+        let mut resp = handle_select(&end_point, req);
+        assert_eq!(row_cnt(resp.get_chunks()), 5);
+        let spliter = ChunkSpliter::new(resp.take_chunks().into_vec());
+        for (row, (id, name, cnt)) in spliter.zip(exp) {
+            let name_datum = name.map(|s| s.as_bytes()).into();
+            let expected_encoded =
+                datum::encode_value(&[(id as i64).into(), name_datum, (cnt as i64).into()])
+                    .unwrap();
+            assert_eq!(id as i64, row.handle);
+            assert_eq!(row.data, &*expected_encoded);
+        }
     }
     end_point.stop().unwrap().join().unwrap();
 }
@@ -931,15 +1331,32 @@ fn test_order_by_pk_with_select_from_index() {
 
     let product = ProductTable::new();
     let (_, mut end_point) = init_with_data(&product, &data);
-    let req = Select::from_index(&product.table, product.name)
-        .order_by(product.id, true)
-        .limit(5)
-        .build();
-    let mut resp = handle_select(&end_point, req);
-    assert_eq!(row_cnt(resp.get_chunks()), 5);
-    let spliter = ChunkSpliter::new(resp.take_chunks().into_vec());
-    for (row, (id, _, _)) in spliter.zip(data.drain(..5)) {
-        assert_eq!(id, row.handle);
+    let expect: Vec<_> = data.drain(..5).collect();
+    {
+        // for selection
+        let req = Select::from_index(&product.table, product.name)
+            .order_by(product.id, true)
+            .limit(5)
+            .build();
+        let mut resp = handle_select(&end_point, req);
+        assert_eq!(row_cnt(resp.get_chunks()), 5);
+        let spliter = ChunkSpliter::new(resp.take_chunks().into_vec());
+        for (row, (id, _, _)) in spliter.zip(expect.clone()) {
+            assert_eq!(id, row.handle);
+        }
+    }
+    {
+        // for dag
+        let req = DAGSelect::from_index(&product.table, product.name)
+            .order_by(product.id, true)
+            .limit(5)
+            .build();
+        let mut resp = handle_select(&end_point, req);
+        assert_eq!(row_cnt(resp.get_chunks()), 5);
+        let spliter = ChunkSpliter::new(resp.take_chunks().into_vec());
+        for (row, (id, _, _)) in spliter.zip(expect) {
+            assert_eq!(id, row.handle);
+        }
     }
     end_point.stop().unwrap().join().unwrap();
 }
@@ -957,16 +1374,34 @@ fn test_limit() {
 
     let product = ProductTable::new();
     let (_, mut end_point) = init_with_data(&product, &data);
-
-    let req = Select::from(&product.table).limit(5).build();
-    let mut resp = handle_select(&end_point, req);
-    assert_eq!(row_cnt(resp.get_chunks()), 5);
-    let spliter = ChunkSpliter::new(resp.take_chunks().into_vec());
-    for (row, (id, name, cnt)) in spliter.zip(data.drain(..5)) {
-        let name_datum = name.map(|s| s.as_bytes()).into();
-        let expected_encoded = datum::encode_value(&[id.into(), name_datum, cnt.into()]).unwrap();
-        assert_eq!(id, row.handle);
-        assert_eq!(row.data, &*expected_encoded);
+    let expect: Vec<_> = data.drain(..5).collect();
+    {
+        // for selection
+        let req = Select::from(&product.table).limit(5).build();
+        let mut resp = handle_select(&end_point, req);
+        assert_eq!(row_cnt(resp.get_chunks()), 5);
+        let spliter = ChunkSpliter::new(resp.take_chunks().into_vec());
+        for (row, (id, name, cnt)) in spliter.zip(expect.clone()) {
+            let name_datum = name.map(|s| s.as_bytes()).into();
+            let expected_encoded = datum::encode_value(&[id.into(), name_datum, cnt.into()])
+                .unwrap();
+            assert_eq!(id, row.handle);
+            assert_eq!(row.data, &*expected_encoded);
+        }
+    }
+    {
+        // for dag
+        let req = DAGSelect::from(&product.table).limit(5).build();
+        let mut resp = handle_select(&end_point, req);
+        assert_eq!(row_cnt(resp.get_chunks()), 5);
+        let spliter = ChunkSpliter::new(resp.take_chunks().into_vec());
+        for (row, (id, name, cnt)) in spliter.zip(expect) {
+            let name_datum = name.map(|s| s.as_bytes()).into();
+            let expected_encoded = datum::encode_value(&[id.into(), name_datum, cnt.into()])
+                .unwrap();
+            assert_eq!(id, row.handle);
+            assert_eq!(row.data, &*expected_encoded);
+        }
     }
 
     end_point.stop().unwrap().join().unwrap();
@@ -985,17 +1420,38 @@ fn test_reverse() {
 
     let product = ProductTable::new();
     let (_, mut end_point) = init_with_data(&product, &data);
-
-    let req = Select::from(&product.table).limit(5).order_by_pk(true).build();
-    let mut resp = handle_select(&end_point, req);
-    assert_eq!(row_cnt(resp.get_chunks()), 5);
-    let spliter = ChunkSpliter::new(resp.take_chunks().into_vec());
     data.reverse();
-    for (row, (id, name, cnt)) in spliter.zip(data.drain(..5)) {
-        let name_datum = name.map(|s| s.as_bytes()).into();
-        let expected_encoded = datum::encode_value(&[id.into(), name_datum, cnt.into()]).unwrap();
-        assert_eq!(id, row.handle);
-        assert_eq!(row.data, &*expected_encoded);
+    let expect: Vec<_> = data.drain(..5).collect();
+
+    {
+        // for selection
+        let req = Select::from(&product.table).limit(5).order_by_pk(true).build();
+        let mut resp = handle_select(&end_point, req);
+        assert_eq!(row_cnt(resp.get_chunks()), 5);
+        let spliter = ChunkSpliter::new(resp.take_chunks().into_vec());
+
+        for (row, (id, name, cnt)) in spliter.zip(expect.clone()) {
+            let name_datum = name.map(|s| s.as_bytes()).into();
+            let expected_encoded = datum::encode_value(&[id.into(), name_datum, cnt.into()])
+                .unwrap();
+            assert_eq!(id, row.handle);
+            assert_eq!(row.data, &*expected_encoded);
+        }
+    }
+
+    {
+        // for dag
+        let req = DAGSelect::from(&product.table).limit(5).order_by(product.id, true).build();
+        let mut resp = handle_select(&end_point, req);
+        assert_eq!(row_cnt(resp.get_chunks()), 5);
+        let spliter = ChunkSpliter::new(resp.take_chunks().into_vec());
+        for (row, (id, name, cnt)) in spliter.zip(expect) {
+            let name_datum = name.map(|s| s.as_bytes()).into();
+            let expected_encoded = datum::encode_value(&[id.into(), name_datum, cnt.into()])
+                .unwrap();
+            assert_eq!(id, row.handle);
+            assert_eq!(row.data, &*expected_encoded);
+        }
     }
 
     end_point.stop().unwrap().join().unwrap();
@@ -1025,15 +1481,30 @@ fn test_index() {
 
     let product = ProductTable::new();
     let (_, mut end_point) = init_with_data(&product, &data);
+    {
+        // for selection
+        let req = Select::from_index(&product.table, product.id).build();
+        let mut resp = handle_select(&end_point, req);
+        assert_eq!(row_cnt(resp.get_chunks()), data.len());
+        let spliter = ChunkSpliter::new(resp.take_chunks().into_vec());
+        let mut handles: Vec<_> = spliter.map(|row| row.handle).collect();
+        handles.sort();
+        for (&h, (id, _, _)) in handles.iter().zip(data.clone()) {
+            assert_eq!(id, h);
+        }
+    }
 
-    let req = Select::from_index(&product.table, product.id).build();
-    let mut resp = handle_select(&end_point, req);
-    assert_eq!(row_cnt(resp.get_chunks()), data.len());
-    let spliter = ChunkSpliter::new(resp.take_chunks().into_vec());
-    let mut handles: Vec<_> = spliter.map(|row| row.handle).collect();
-    handles.sort();
-    for (&h, (id, _, _)) in handles.iter().zip(data) {
-        assert_eq!(id, h);
+    {
+        // for dag
+        let req = DAGSelect::from_index(&product.table, product.id).build();
+        let mut resp = handle_select(&end_point, req);
+        assert_eq!(row_cnt(resp.get_chunks()), data.len());
+        let spliter = ChunkSpliter::new(resp.take_chunks().into_vec());
+        let mut handles: Vec<_> = spliter.map(|row| row.handle).collect();
+        handles.sort();
+        for (&h, (id, _, _)) in handles.iter().zip(data) {
+            assert_eq!(id, h);
+        }
     }
 
     end_point.stop().unwrap().join().unwrap();
@@ -1052,15 +1523,34 @@ fn test_index_reverse_limit() {
 
     let product = ProductTable::new();
     let (_, mut end_point) = init_with_data(&product, &data);
-
-    let req = Select::from_index(&product.table, product.id).limit(5).order_by_pk(true).build();
-    let mut resp = handle_select(&end_point, req);
-    assert_eq!(row_cnt(resp.get_chunks()), 5);
-    let spliter = ChunkSpliter::new(resp.take_chunks().into_vec());
-    let handles = spliter.map(|row| row.handle);
     data.reverse();
-    for (h, (id, _, _)) in handles.zip(data.drain(..5)) {
-        assert_eq!(id, h);
+    let expect: Vec<_> = data.drain(..5).collect();
+    {
+        // selection
+        let req = Select::from_index(&product.table, product.id).limit(5).order_by_pk(true).build();
+        let mut resp = handle_select(&end_point, req);
+        assert_eq!(row_cnt(resp.get_chunks()), 5);
+        let spliter = ChunkSpliter::new(resp.take_chunks().into_vec());
+        let handles = spliter.map(|row| row.handle);
+        for (h, (id, _, _)) in handles.zip(expect.clone()) {
+            assert_eq!(id, h);
+        }
+    }
+    {
+        // for dag
+        let req = DAGSelect::from_index(&product.table, product.id)
+            .limit(5)
+            .order_by(product.id, true)
+            .build();
+
+        let mut resp = handle_select(&end_point, req);
+        assert_eq!(row_cnt(resp.get_chunks()), 5);
+        let spliter = ChunkSpliter::new(resp.take_chunks().into_vec());
+        let handles = spliter.map(|row| row.handle);
+        data.reverse();
+        for (h, (id, _, _)) in handles.zip(expect) {
+            assert_eq!(id, h);
+        }
     }
 
     end_point.stop().unwrap().join().unwrap();
@@ -1079,17 +1569,30 @@ fn test_limit_oom() {
 
     let product = ProductTable::new();
     let (_, mut end_point) = init_with_data(&product, &data);
-
-    let req = Select::from_index(&product.table, product.id).limit(100000000).build();
-    let mut resp = handle_select(&end_point, req);
-    assert_eq!(row_cnt(resp.get_chunks()), data.len());
-    let spliter = ChunkSpliter::new(resp.take_chunks().into_vec());
-    let mut handles: Vec<_> = spliter.map(|row| row.handle).collect();
-    handles.sort();
-    for (&h, (id, _, _)) in handles.iter().zip(data) {
-        assert_eq!(id, h);
+    {
+        // for selection
+        let req = Select::from_index(&product.table, product.id).limit(100000000).build();
+        let mut resp = handle_select(&end_point, req);
+        assert_eq!(row_cnt(resp.get_chunks()), data.len());
+        let spliter = ChunkSpliter::new(resp.take_chunks().into_vec());
+        let mut handles: Vec<_> = spliter.map(|row| row.handle).collect();
+        handles.sort();
+        for (&h, (id, _, _)) in handles.iter().zip(data.clone()) {
+            assert_eq!(id, h);
+        }
     }
-
+    {
+        // for dag
+        let req = DAGSelect::from_index(&product.table, product.id).limit(100000000).build();
+        let mut resp = handle_select(&end_point, req);
+        assert_eq!(row_cnt(resp.get_chunks()), data.len());
+        let spliter = ChunkSpliter::new(resp.take_chunks().into_vec());
+        let mut handles: Vec<_> = spliter.map(|row| row.handle).collect();
+        handles.sort();
+        for (&h, (id, _, _)) in handles.iter().zip(data) {
+            assert_eq!(id, h);
+        }
+    }
     end_point.stop().unwrap().join().unwrap();
 }
 
@@ -1112,10 +1615,19 @@ fn test_del_select() {
     let name_datum = name.map(|s| s.as_bytes()).into();
     store.delete_from(&product.table).execute(id, vec![id.into(), name_datum, cnt.into()]);
     store.commit();
+    {
+        // for selection
+        let req = Select::from_index(&product.table, product.id).build();
+        let resp = handle_select(&end_point, req);
+        assert_eq!(row_cnt(resp.get_chunks()), data.len());
+    }
+    {
+        // for dag
+        let req = DAGSelect::from_index(&product.table, product.id).build();
+        let resp = handle_select(&end_point, req);
+        assert_eq!(row_cnt(resp.get_chunks()), data.len());
 
-    let req = Select::from_index(&product.table, product.id).build();
-    let resp = handle_select(&end_point, req);
-    assert_eq!(row_cnt(resp.get_chunks()), data.len());
+    }
 
     end_point.stop().unwrap().join().unwrap();
 }
@@ -1131,16 +1643,32 @@ fn test_index_group_by() {
 
     let product = ProductTable::new();
     let (_, mut end_point) = init_with_data(&product, &data);
-
-    let req = Select::from_index(&product.table, product.name).group_by(&[product.name]).build();
-    let mut resp = handle_select(&end_point, req);
-    // should only have name:0, name:2 and name:1
-    assert_eq!(row_cnt(resp.get_chunks()), 3);
-    let spliter = ChunkSpliter::new(resp.take_chunks().into_vec());
-    for (row, name) in spliter.zip(&[b"name:0", b"name:1", b"name:2"]) {
-        let gk = datum::encode_value(&[Datum::Bytes(name.to_vec())]).unwrap();
-        let expected_encoded = datum::encode_value(&[Datum::Bytes(gk)]).unwrap();
-        assert_eq!(row.data, &*expected_encoded);
+    {
+        // for selection
+        let req =
+            Select::from_index(&product.table, product.name).group_by(&[product.name]).build();
+        let mut resp = handle_select(&end_point, req);
+        // should only have name:0, name:2 and name:1
+        assert_eq!(row_cnt(resp.get_chunks()), 3);
+        let spliter = ChunkSpliter::new(resp.take_chunks().into_vec());
+        for (row, name) in spliter.zip(&[b"name:0", b"name:1", b"name:2"]) {
+            let gk = datum::encode_value(&[Datum::Bytes(name.to_vec())]).unwrap();
+            let expected_encoded = datum::encode_value(&[Datum::Bytes(gk)]).unwrap();
+            assert_eq!(row.data, &*expected_encoded);
+        }
+    }
+    {
+        // for dag
+        let req =
+            DAGSelect::from_index(&product.table, product.name).group_by(&[product.name]).build();
+        let mut resp = handle_select(&end_point, req);
+        // should only have name:0, name:2 and name:1
+        assert_eq!(row_cnt(resp.get_chunks()), 3);
+        let spliter = ChunkSpliter::new(resp.take_chunks().into_vec());
+        for (row, name) in spliter.zip(&[b"name:0", b"name:1", b"name:2"]) {
+            let expected_encoded = datum::encode_value(&[Datum::Bytes(name.to_vec())]).unwrap();
+            assert_eq!(row.data, &*expected_encoded);
+        }
     }
 
     end_point.stop().unwrap().join().unwrap();
@@ -1159,7 +1687,7 @@ fn test_index_aggr_count() {
 
     let product = ProductTable::new();
     let (_, mut end_point) = init_with_data(&product, &data);
-
+    // for selection
     let req = Select::from_index(&product.table, product.name).count().build();
     let mut resp = handle_select(&end_point, req);
     assert_eq!(row_cnt(resp.get_chunks()), 1);
@@ -1168,24 +1696,52 @@ fn test_index_aggr_count() {
     let mut expected_encoded = datum::encode_value(&[gk, Datum::U64(data.len() as u64)]).unwrap();
     assert_eq!(spliter.next().unwrap().data, &*expected_encoded);
 
+    // for dag
+    let req = DAGSelect::from_index(&product.table, product.name).count().build();
+    let mut resp = handle_select(&end_point, req);
+    assert_eq!(row_cnt(resp.get_chunks()), 1);
+    let mut spliter = ChunkSpliter::new(resp.take_chunks().into_vec());
+    let gk = Datum::Bytes(coprocessor::SINGLE_GROUP.to_vec());
+    expected_encoded = datum::encode_value(&[Datum::U64(data.len() as u64), gk]).unwrap();
+    assert_eq!(spliter.next().unwrap().data, &*expected_encoded);
+
+
     let exp = vec![
         (Datum::Null, 1),
         (Datum::Bytes(b"name:0".to_vec()), 2),
         (Datum::Bytes(b"name:3".to_vec()), 1),
         (Datum::Bytes(b"name:5".to_vec()), 2),
     ];
-    let req = Select::from_index(&product.table, product.name)
-        .count()
-        .group_by(&[product.name])
-        .build();
-    resp = handle_select(&end_point, req);
-    assert_eq!(row_cnt(resp.get_chunks()), exp.len());
-    let spliter = ChunkSpliter::new(resp.take_chunks().into_vec());
-    for (row, (name, cnt)) in spliter.zip(exp) {
-        let gk = datum::encode_value(&[name]);
-        let expected_datum = vec![Datum::Bytes(gk.unwrap()), Datum::U64(cnt)];
-        expected_encoded = datum::encode_value(&expected_datum).unwrap();
-        assert_eq!(row.data, &*expected_encoded);
+    {
+        // for selection
+        let req = Select::from_index(&product.table, product.name)
+            .count()
+            .group_by(&[product.name])
+            .build();
+        resp = handle_select(&end_point, req);
+        assert_eq!(row_cnt(resp.get_chunks()), exp.len());
+        let spliter = ChunkSpliter::new(resp.take_chunks().into_vec());
+        for (row, (name, cnt)) in spliter.zip(exp.clone()) {
+            let gk = datum::encode_value(&[name]);
+            let expected_datum = vec![Datum::Bytes(gk.unwrap()), Datum::U64(cnt)];
+            expected_encoded = datum::encode_value(&expected_datum).unwrap();
+            assert_eq!(row.data, &*expected_encoded);
+        }
+    }
+    {
+        // for dag
+        let req = DAGSelect::from_index(&product.table, product.name)
+            .count()
+            .group_by(&[product.name])
+            .build();
+        resp = handle_select(&end_point, req);
+        assert_eq!(row_cnt(resp.get_chunks()), exp.len());
+        let spliter = ChunkSpliter::new(resp.take_chunks().into_vec());
+        for (row, (name, cnt)) in spliter.zip(exp) {
+            let expected_datum = vec![Datum::U64(cnt), name];
+            expected_encoded = datum::encode_value(&expected_datum).unwrap();
+            assert_eq!(row.data, &*expected_encoded);
+        }
     }
 
     let exp = vec![
@@ -1195,18 +1751,37 @@ fn test_index_aggr_count() {
         (vec![Datum::Bytes(b"name:3".to_vec()), Datum::I64(3)], 1),
         (vec![Datum::Bytes(b"name:5".to_vec()), Datum::I64(4)], 2),
     ];
-    let req = Select::from_index(&product.table, product.name)
-        .count()
-        .group_by(&[product.name, product.count])
-        .build();
-    resp = handle_select(&end_point, req);
-    assert_eq!(row_cnt(resp.get_chunks()), exp.len());
-    let spliter = ChunkSpliter::new(resp.take_chunks().into_vec());
-    for (row, (gk_data, cnt)) in spliter.zip(exp) {
-        let gk = datum::encode_value(&gk_data);
-        let expected_datum = vec![Datum::Bytes(gk.unwrap()), Datum::U64(cnt)];
-        expected_encoded = datum::encode_value(&expected_datum).unwrap();
-        assert_eq!(row.data, &*expected_encoded);
+    {
+        // for selection
+        let req = Select::from_index(&product.table, product.name)
+            .count()
+            .group_by(&[product.name, product.count])
+            .build();
+        resp = handle_select(&end_point, req);
+        assert_eq!(row_cnt(resp.get_chunks()), exp.len());
+        let spliter = ChunkSpliter::new(resp.take_chunks().into_vec());
+        for (row, (gk_data, cnt)) in spliter.zip(exp.clone()) {
+            let gk = datum::encode_value(&gk_data);
+            let expected_datum = vec![Datum::Bytes(gk.unwrap()), Datum::U64(cnt)];
+            expected_encoded = datum::encode_value(&expected_datum).unwrap();
+            assert_eq!(row.data, &*expected_encoded);
+        }
+    }
+    {
+        // for dag
+        let req = DAGSelect::from_index(&product.table, product.name)
+            .count()
+            .group_by(&[product.name, product.count])
+            .build();
+        resp = handle_select(&end_point, req);
+        assert_eq!(row_cnt(resp.get_chunks()), exp.len());
+        let spliter = ChunkSpliter::new(resp.take_chunks().into_vec());
+        for (row, (gk_data, cnt)) in spliter.zip(exp) {
+            let mut expected_datum = vec![Datum::U64(cnt)];
+            expected_datum.extend_from_slice(gk_data.as_slice());
+            expected_encoded = datum::encode_value(&expected_datum).unwrap();
+            assert_eq!(row.data, &*expected_encoded);
+        }
     }
 
     end_point.stop().unwrap().join().unwrap();
@@ -1232,18 +1807,37 @@ fn test_index_aggr_first() {
         (Datum::Bytes(b"name:3".to_vec()), 2),
         (Datum::Bytes(b"name:5".to_vec()), 5),
     ];
-    let req = Select::from_index(&product.table, product.name)
-        .first(product.id)
-        .group_by(&[product.name])
-        .build();
-    let mut resp = handle_select(&end_point, req);
-    assert_eq!(row_cnt(resp.get_chunks()), exp.len());
-    let spliter = ChunkSpliter::new(resp.take_chunks().into_vec());
-    for (row, (name, id)) in spliter.zip(exp) {
-        let gk = datum::encode_value(&[name]).unwrap();
-        let expected_datum = vec![Datum::Bytes(gk), Datum::I64(id)];
-        let expected_encoded = datum::encode_value(&expected_datum).unwrap();
-        assert_eq!(row.data, &*expected_encoded);
+    {
+        // for selection
+        let req = Select::from_index(&product.table, product.name)
+            .first(product.id)
+            .group_by(&[product.name])
+            .build();
+        let mut resp = handle_select(&end_point, req);
+        assert_eq!(row_cnt(resp.get_chunks()), exp.len());
+        let spliter = ChunkSpliter::new(resp.take_chunks().into_vec());
+        for (row, (name, id)) in spliter.zip(exp.clone()) {
+            let gk = datum::encode_value(&[name]).unwrap();
+            let expected_datum = vec![Datum::Bytes(gk), Datum::I64(id)];
+            let expected_encoded = datum::encode_value(&expected_datum).unwrap();
+            assert_eq!(row.data, &*expected_encoded);
+        }
+    }
+    {
+        // for dag
+        let req = DAGSelect::from_index(&product.table, product.name)
+            .first(product.id)
+            .group_by(&[product.name])
+            .build();
+        let mut resp = handle_select(&end_point, req);
+        assert_eq!(row_cnt(resp.get_chunks()), exp.len());
+        let spliter = ChunkSpliter::new(resp.take_chunks().into_vec());
+        for (row, (name, id)) in spliter.zip(exp) {
+            let expected_datum = vec![Datum::I64(id), name];
+            let expected_encoded = datum::encode_value(&expected_datum).unwrap();
+            assert_eq!(row.data, &*expected_encoded);
+        }
+
     }
 
     end_point.stop().unwrap().join().unwrap();
@@ -1276,18 +1870,36 @@ fn test_index_aggr_avg() {
                    (Datum::Bytes(b"name:3".to_vec()), (Datum::Dec(3.into()), 1)),
                    (Datum::Bytes(b"name:4".to_vec()), (Datum::Null, 0)),
                    (Datum::Bytes(b"name:5".to_vec()), (Datum::Dec(8.into()), 2))];
-    let req = Select::from_index(&product.table, product.name)
-        .avg(product.count)
-        .group_by(&[product.name])
-        .build();
-    let mut resp = handle_select(&end_point, req);
-    assert_eq!(row_cnt(resp.get_chunks()), exp.len());
-    let spliter = ChunkSpliter::new(resp.take_chunks().into_vec());
-    for (row, (name, (sum, cnt))) in spliter.zip(exp) {
-        let gk = datum::encode_value(&[name]).unwrap();
-        let expected_datum = vec![Datum::Bytes(gk), Datum::U64(cnt), sum];
-        let expected_encoded = datum::encode_value(&expected_datum).unwrap();
-        assert_eq!(row.data, &*expected_encoded);
+
+    {
+        let req = Select::from_index(&product.table, product.name)
+            .avg(product.count)
+            .group_by(&[product.name])
+            .build();
+        let mut resp = handle_select(&end_point, req);
+        assert_eq!(row_cnt(resp.get_chunks()), exp.len());
+        let spliter = ChunkSpliter::new(resp.take_chunks().into_vec());
+        for (row, (name, (sum, cnt))) in spliter.zip(exp.clone()) {
+            let gk = datum::encode_value(&[name]).unwrap();
+            let expected_datum = vec![Datum::Bytes(gk), Datum::U64(cnt), sum];
+            let expected_encoded = datum::encode_value(&expected_datum).unwrap();
+            assert_eq!(row.data, &*expected_encoded);
+        }
+    }
+    {
+        // for dag
+        let req = DAGSelect::from_index(&product.table, product.name)
+            .avg(product.count)
+            .group_by(&[product.name])
+            .build();
+        let mut resp = handle_select(&end_point, req);
+        assert_eq!(row_cnt(resp.get_chunks()), exp.len());
+        let spliter = ChunkSpliter::new(resp.take_chunks().into_vec());
+        for (row, (name, (sum, cnt))) in spliter.zip(exp) {
+            let expected_datum = vec![Datum::U64(cnt), sum, name];
+            let expected_encoded = datum::encode_value(&expected_datum).unwrap();
+            assert_eq!(row.data, &*expected_encoded);
+        }
     }
     end_point.stop().unwrap();
 }
@@ -1312,18 +1924,36 @@ fn test_index_aggr_sum() {
         (Datum::Bytes(b"name:3".to_vec()), 3),
         (Datum::Bytes(b"name:5".to_vec()), 8),
     ];
-    let req = Select::from_index(&product.table, product.name)
-        .sum(product.count)
-        .group_by(&[product.name])
-        .build();
-    let mut resp = handle_select(&end_point, req);
-    assert_eq!(row_cnt(resp.get_chunks()), exp.len());
-    let spliter = ChunkSpliter::new(resp.take_chunks().into_vec());
-    for (row, (name, cnt)) in spliter.zip(exp) {
-        let gk = datum::encode_value(&[name]).unwrap();
-        let expected_datum = vec![Datum::Bytes(gk), Datum::Dec(cnt.into())];
-        let expected_encoded = datum::encode_value(&expected_datum).unwrap();
-        assert_eq!(row.data, &*expected_encoded);
+    {
+        // for selection
+        let req = Select::from_index(&product.table, product.name)
+            .sum(product.count)
+            .group_by(&[product.name])
+            .build();
+        let mut resp = handle_select(&end_point, req);
+        assert_eq!(row_cnt(resp.get_chunks()), exp.len());
+        let spliter = ChunkSpliter::new(resp.take_chunks().into_vec());
+        for (row, (name, cnt)) in spliter.zip(exp.clone()) {
+            let gk = datum::encode_value(&[name]).unwrap();
+            let expected_datum = vec![Datum::Bytes(gk), Datum::Dec(cnt.into())];
+            let expected_encoded = datum::encode_value(&expected_datum).unwrap();
+            assert_eq!(row.data, &*expected_encoded);
+        }
+    }
+    {
+        // for dag
+        let req = DAGSelect::from_index(&product.table, product.name)
+            .sum(product.count)
+            .group_by(&[product.name])
+            .build();
+        let mut resp = handle_select(&end_point, req);
+        assert_eq!(row_cnt(resp.get_chunks()), exp.len());
+        let spliter = ChunkSpliter::new(resp.take_chunks().into_vec());
+        for (row, (name, cnt)) in spliter.zip(exp) {
+            let expected_datum = vec![Datum::Dec(cnt.into()), name];
+            let expected_encoded = datum::encode_value(&expected_datum).unwrap();
+            assert_eq!(row.data, &*expected_encoded);
+        }
     }
     end_point.stop().unwrap();
 }
@@ -1359,19 +1989,38 @@ fn test_index_aggr_extre() {
         (Datum::Bytes(b"name:5".to_vec()), Datum::I64(5), Datum::I64(4)),
         (Datum::Bytes(b"name:6".to_vec()), Datum::Null, Datum::Null),
     ];
-    let req = Select::from_index(&product.table, product.name)
-        .max(product.count)
-        .min(product.count)
-        .group_by(&[product.name])
-        .build();
-    let mut resp = handle_select(&end_point, req);
-    assert_eq!(row_cnt(resp.get_chunks()), exp.len());
-    let spliter = ChunkSpliter::new(resp.take_chunks().into_vec());
-    for (row, (name, max, min)) in spliter.zip(exp) {
-        let gk = datum::encode_value(&[name]).unwrap();
-        let expected_datum = vec![Datum::Bytes(gk), max, min];
-        let expected_encoded = datum::encode_value(&expected_datum).unwrap();
-        assert_eq!(row.data, &*expected_encoded);
+    {
+        // for selection
+        let req = Select::from_index(&product.table, product.name)
+            .max(product.count)
+            .min(product.count)
+            .group_by(&[product.name])
+            .build();
+        let mut resp = handle_select(&end_point, req);
+        assert_eq!(row_cnt(resp.get_chunks()), exp.len());
+        let spliter = ChunkSpliter::new(resp.take_chunks().into_vec());
+        for (row, (name, max, min)) in spliter.zip(exp.clone()) {
+            let gk = datum::encode_value(&[name]).unwrap();
+            let expected_datum = vec![Datum::Bytes(gk), max, min];
+            let expected_encoded = datum::encode_value(&expected_datum).unwrap();
+            assert_eq!(row.data, &*expected_encoded);
+        }
+    }
+    {
+        // for dag
+        let req = DAGSelect::from_index(&product.table, product.name)
+            .max(product.count)
+            .min(product.count)
+            .group_by(&[product.name])
+            .build();
+        let mut resp = handle_select(&end_point, req);
+        assert_eq!(row_cnt(resp.get_chunks()), exp.len());
+        let spliter = ChunkSpliter::new(resp.take_chunks().into_vec());
+        for (row, (name, max, min)) in spliter.zip(exp) {
+            let expected_datum = vec![max, min, name];
+            let expected_encoded = datum::encode_value(&expected_datum).unwrap();
+            assert_eq!(row.data, &*expected_encoded);
+        }
     }
     end_point.stop().unwrap();
 }
@@ -1405,6 +2054,50 @@ fn test_where() {
     };
 
     let req = Select::from(&product.table).where_expr(cond).build();
+    let mut resp = handle_select(&end_point, req);
+    assert_eq!(row_cnt(resp.get_chunks()), 1);
+    let mut spliter = ChunkSpliter::new(resp.take_chunks().into_vec());
+    let row = spliter.next().unwrap();
+    let (id, name, cnt) = data[2];
+    let name_datum = name.map(|s| s.as_bytes()).into();
+    let expected_encoded = datum::encode_value(&[Datum::I64(id), name_datum, cnt.into()]).unwrap();
+    assert_eq!(id, row.handle);
+    assert_eq!(row.data, &*expected_encoded);
+
+    end_point.stop().unwrap().join().unwrap();
+}
+
+
+#[test]
+fn test_where_for_dag() {
+    let data = vec![
+        (1, Some("name:0"), 2),
+        (2, Some("name:4"), 3),
+        (4, Some("name:3"), 1),
+        (5, Some("name:1"), 4),
+    ];
+
+    let product = ProductTable::new();
+    let (_, mut end_point) = init_with_data(&product, &data);
+    let cols = product.table.get_table_columns();
+    let cond = {
+        let mut col = Expr::new();
+        col.set_tp(ExprType::ColumnRef);
+        let count_offset = offset_for_column(&cols, product.count.id);
+        col.mut_val().encode_i64(count_offset).unwrap();
+
+        let mut value = Expr::new();
+        value.set_tp(ExprType::String);
+        value.set_val(String::from("2").into_bytes());
+
+        let mut cond = Expr::new();
+        cond.set_tp(ExprType::LT);
+        cond.mut_children().push(col);
+        cond.mut_children().push(value);
+        cond
+    };
+
+    let req = DAGSelect::from(&product.table).where_expr(cond).build();
     let mut resp = handle_select(&end_point, req);
     assert_eq!(row_cnt(resp.get_chunks()), 1);
     let mut spliter = ChunkSpliter::new(resp.take_chunks().into_vec());
@@ -1506,6 +2199,96 @@ fn test_handle_truncate() {
 }
 
 #[test]
+fn test_handle_truncate_for_dag() {
+    let data = vec![
+        (1, Some("name:0"), 2),
+        (2, Some("name:4"), 3),
+        (4, Some("name:3"), 1),
+        (5, Some("name:1"), 4),
+    ];
+
+    let product = ProductTable::new();
+    let (_, mut end_point) = init_with_data(&product, &data);
+    let cols = product.table.get_table_columns();
+    let cases = vec![{
+                         // count > "2x"
+                         let mut col = Expr::new();
+                         col.set_tp(ExprType::ColumnRef);
+                         let count_offset = offset_for_column(&cols, product.count.id);
+                         col.mut_val().encode_i64(count_offset).unwrap();
+
+                         // "2x" will be truncated.
+                         let mut value = Expr::new();
+                         value.set_tp(ExprType::String);
+                         value.set_val(String::from("2x").into_bytes());
+
+                         let mut cond = Expr::new();
+                         cond.set_tp(ExprType::LT);
+                         cond.mut_children().push(col);
+                         cond.mut_children().push(value);
+                         cond
+                     },
+                     {
+                         // id
+                         let mut col_id = Expr::new();
+                         col_id.set_tp(ExprType::ColumnRef);
+                         let id_offset = offset_for_column(&cols, product.id.id);
+                         col_id.mut_val().encode_i64(id_offset).unwrap();
+
+                         // "3x" will be truncated.
+                         let mut value = Expr::new();
+                         value.set_tp(ExprType::String);
+                         value.set_val(String::from("3x").into_bytes());
+
+                         // count
+                         let mut col_count = Expr::new();
+                         col_count.set_tp(ExprType::ColumnRef);
+                         let count_offset = offset_for_column(&cols, product.count.id);
+                         col_count.mut_val().encode_i64(count_offset).unwrap();
+
+                         // "3x" + count
+                         let mut plus = Expr::new();
+                         plus.set_tp(ExprType::Plus);
+                         plus.mut_children().push(value);
+                         plus.mut_children().push(col_count);
+
+                         // id = "3x" + count
+                         let mut cond = Expr::new();
+                         cond.set_tp(ExprType::EQ);
+                         cond.mut_children().push(col_id);
+                         cond.mut_children().push(plus);
+                         cond
+                     }];
+
+    for cond in cases {
+        // Ignore truncate error.
+        let req = DAGSelect::from(&product.table)
+            .where_expr(cond.clone())
+            .build_with(&[FLAG_IGNORE_TRUNCATE]);
+        let mut resp = handle_select(&end_point, req);
+        assert_eq!(row_cnt(resp.get_chunks()), 1);
+        let mut spliter = ChunkSpliter::new(resp.take_chunks().into_vec());
+        let row = spliter.next().unwrap();
+        let (id, name, cnt) = data[2];
+        let name_datum = name.map(|s| s.as_bytes()).into();
+        let expected_encoded = datum::encode_value(&[Datum::I64(id), name_datum, cnt.into()])
+            .unwrap();
+        assert_eq!(id, row.handle);
+        assert_eq!(row.data, &*expected_encoded);
+
+        // Do NOT ignore truncate error.
+        let req = DAGSelect::from(&product.table).where_expr(cond.clone()).build();
+        let (tx, rx) = mpsc::channel();
+        let req = RequestTask::new(req, box move |r| tx.send(r).unwrap());
+        end_point.schedule(EndPointTask::Request(req)).unwrap();
+        let resp = rx.recv().unwrap();
+        assert!(!resp.get_other_error().is_empty());
+    }
+
+    end_point.stop().unwrap().join().unwrap();
+}
+
+#[test]
 fn test_default_val() {
     let mut data = vec![
         (1, Some("name:0"), 2),
@@ -1527,17 +2310,34 @@ fn test_default_val() {
     tbl.id = product.table.id;
 
     let (_, mut end_point) = init_with_data(&product, &data);
-
-    let req = Select::from(&tbl).limit(5).build();
-    let mut resp = handle_select(&end_point, req);
-    assert_eq!(row_cnt(resp.get_chunks()), 5);
-    let spliter = ChunkSpliter::new(resp.take_chunks().into_vec());
-    for (row, (id, name, cnt)) in spliter.zip(data.drain(..5)) {
-        let name_datum = name.map(|s| s.as_bytes()).into();
-        let expected_encoded =
-            datum::encode_value(&[id.into(), name_datum, cnt.into(), Datum::I64(3)]).unwrap();
-        assert_eq!(id, row.handle);
-        assert_eq!(row.data, &*expected_encoded);
+    let expect: Vec<_> = data.drain(..5).collect();
+    {
+        // for selection
+        let req = Select::from(&tbl).limit(5).build();
+        let mut resp = handle_select(&end_point, req);
+        assert_eq!(row_cnt(resp.get_chunks()), 5);
+        let spliter = ChunkSpliter::new(resp.take_chunks().into_vec());
+        for (row, (id, name, cnt)) in spliter.zip(expect.clone()) {
+            let name_datum = name.map(|s| s.as_bytes()).into();
+            let expected_encoded =
+                datum::encode_value(&[id.into(), name_datum, cnt.into(), Datum::I64(3)]).unwrap();
+            assert_eq!(id, row.handle);
+            assert_eq!(row.data, &*expected_encoded);
+        }
+    }
+    {
+        // for dag
+        let req = DAGSelect::from(&tbl).limit(5).build();
+        let mut resp = handle_select(&end_point, req);
+        assert_eq!(row_cnt(resp.get_chunks()), 5);
+        let spliter = ChunkSpliter::new(resp.take_chunks().into_vec());
+        for (row, (id, name, cnt)) in spliter.zip(expect) {
+            let name_datum = name.map(|s| s.as_bytes()).into();
+            let expected_encoded =
+                datum::encode_value(&[id.into(), name_datum, cnt.into(), Datum::I64(3)]).unwrap();
+            assert_eq!(id, row.handle);
+            assert_eq!(row.data, &*expected_encoded);
+        }
     }
 
     end_point.stop().unwrap().join().unwrap();
